% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_grid.R
\name{diff_grid}
\alias{diff_grid}
\title{Distribute the data}
\usage{
diff_grid(gridmapX, gridmapY, var)
}
\arguments{
\item{input}{Path to the input file}

\item{lx}{Size of}
}
\value{
gridmap
}
\description{
This function ...
}
