% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{Distribute the data over a grid}
\usage{
distribute(
  input,
  lx,
  ly = lx,
  vars,
  bbox,
  crs = "+proj=utm +zone=21 +south +datum=WGS84 +units=m +no_defs",
  compute_distances = TRUE,
  ...
)
}
\arguments{
\item{input}{Object of class \link{sf}.}

\item{lx, ly}{Numeric of length 1, with the size of gridcells.}

\item{vars}{Vector of variables names of data.}

\item{bbox}{Object of class \link{bbox}.}

\item{crs}{Coordinates reference system.}

\item{compute_distances}{Logical, if \code{TRUE} (default) the distance matrix between grid cells is calculated.}

\item{...}{optional}
}
\value{
Object of class \code{gridmap}
}
\description{
\code{distribute} makes a regular grid of cells with sizes \code{c(lx,ly)}, divide the raw data stored in original polygons and distribute it over the grid.
}
\examples{
\dontrun{
data(input)
distributed <- distribute(input = input, lx = 100, vars = c("nivel_edu_alto", "nivel_edu_bajo"))
plot(distributed)
}
}
