% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environments.R
\name{environments}
\alias{environments}
\title{Distribute the data}
\usage{
environments(gridmap, gamma)
}
\arguments{
\item{input}{Path to the input file}
}
\value{
gridmap
}
\description{
This function ...
}
