% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Estimate ...}
\usage{
estimate(
  input,
  vars,
  lx = 100,
  ly,
  Ngrid,
  smoothing = "none",
  sigma,
  graficar = F,
  verbose = T
)
}
\arguments{
\item{input}{Path to the input file}
}
\value{
A data.table
}
\description{
This function ...
}
