% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_profile.R
\name{plot.seg_profile}
\alias{plot.seg_profile}
\title{Distribute the data}
\usage{
\method{plot}{seg_profile}(seg_profile)
}
\arguments{
\item{input}{Path to the input file}
}
\value{
gridmap
}
\description{
This function ...
}
