% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segplot.R
\name{reverselog_trans}
\alias{reverselog_trans}
\title{Estimate ...}
\usage{
reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{Path to the input file}
}
\value{
A data.table
}
\description{
This function ...
}
