% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_density.R
\name{scale_density}
\alias{scale_density}
\title{Estimate ...}
\usage{
scale_density(data, vars)
}
\arguments{
\item{data}{data}

\item{vars}{variables}
}
\value{
A data.table
}
\description{
This function ...
}
