% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_profile.R
\name{seg_profile}
\alias{seg_profile}
\title{Distribute the data}
\usage{
seg_profile(gridmap, gamma, vars, nucleo = "quartic")
}
\arguments{
\item{input}{Path to the input file}
}
\value{
gridmap
}
\description{
This function ...
}
