% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segplot.R
\name{segplot}
\alias{segplot}
\title{Estimate ...}
\usage{
segplot(data, input, var, poligono = T, contour = T, per = 0.9)
}
\arguments{
\item{data}{data}

\item{input}{polygon}

\item{var}{variable}
}
\value{
A data.table
}
\description{
This function ...
}
