% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segsmooth.R
\name{segsmooth}
\alias{segsmooth}
\title{Smooth the data}
\usage{
segsmooth(data, sigma, bleed = F, normalise = F, rescale = T)
}
\arguments{
\item{data}{...}

\item{sigma}{...}
}
\value{
A data.table
}
\description{
This function ...
}
