% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_map}
\alias{simulate_map}
\title{Distribute the data}
\usage{
simulate_map(gridmap, vars, N = 1)
}
\arguments{
\item{vars}{name}

\item{input}{Path to the input file}

\item{lx}{Size of grid (width)}

\item{ly}{Size of grid (height)}

\item{bbox}{square}

\item{crs}{coordinates}

\item{compute_distances}{ar}

\item{...}{optional}
}
\value{
aaaa
}
\description{
This function ...
}
\keyword{internal}
