% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Smooth the data}
\usage{
smooth(data, sigma, bleed = F, normalise = F, rescale = T)
}
\arguments{
\item{data, }{sigma, ...}
}
\value{
A data.table
}
\description{
This function ...
}
