% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothgrid.R
\name{smoothgrid}
\alias{smoothgrid}
\title{Smooth the data}
\usage{
smoothgrid(gridmap, sigma, vars, nucleo = "quartic")
}
\arguments{
\item{sigma}{...}

\item{data}{...}
}
\value{
A data.table
}
\description{
This function ...
}
