% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{Distribute the data}
\usage{
distribute(input, Ngrid, lx, ly, vars, verbose = T)
}
\arguments{
\item{input}{Path to the input file}
}
\value{
A data.table
}
\description{
This function ...
}
