% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{Distribute the data}
\usage{
distribute(
  input,
  lx,
  ly = lx,
  vars,
  bbox,
  crs = "+proj=utm +zone=21 +south +datum=WGS84 +units=m +no_defs",
  compute_distances = TRUE,
  ...
)
}
\arguments{
\item{input}{Path to the input file}

\item{lx}{Size of grid (width)}

\item{ly}{Size of grid (height)}

\item{vars}{name}

\item{bbox}{square}

\item{crs}{coordinates}

\item{compute_distances}{ar}

\item{...}{optional}
}
\value{
gridmap
}
\description{
This function ...
}
